Extract tests for generic return type to separate classes
