+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/condition/ConditionalOnMissingBeanTests.java
@@ -59,7 +59,6 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author Phillip Webb
  * @author Jakub Kubrynski
  * @author Andy Wilkinson
- * @author Uladzislau Seuruk
  */
 @SuppressWarnings("resource")
 class ConditionalOnMissingBeanTests {
@@ -341,70 +340,6 @@ class ConditionalOnMissingBeanTests {
 				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("customExampleBean")));
 	}
 
-	@Test
-	void genericWhenTypeArgumentNotMatches() {
-		this.contextRunner
-				.withUserConfiguration(GenericWithStringTypeArgumentsConfig.class,
-						GenericWithIntegerTypeArgumentsConfig.class)
-				.run((context) -> assertThat(context)
-						.satisfies(exampleBeanRequirement("genericStringExampleBean", "genericIntegerExampleBean")));
-	}
-
-	@Test
-	void genericWhenSubclassTypeArgumentMatches() {
-		this.contextRunner
-				.withUserConfiguration(ParameterizedWithCustomConfig.class, GenericWithStringTypeArgumentsConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("customExampleBean")));
-	}
-
-	@Test
-	void genericWhenSubclassTypeArgumentNotMatches() {
-		this.contextRunner
-				.withUserConfiguration(ParameterizedWithCustomConfig.class, GenericWithIntegerTypeArgumentsConfig.class)
-				.run((context) -> assertThat(context)
-						.satisfies(exampleBeanRequirement("customExampleBean", "genericIntegerExampleBean")));
-	}
-
-	@Test
-	void genericWhenTypeArgumentWithValueMatches() {
-		this.contextRunner
-				.withUserConfiguration(GenericWithStringTypeArgumentsConfig.class,
-						TypeArgumentsConditionWithValueConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("genericStringExampleBean")));
-	}
-
-	@Test
-	void genericWithValueWhenSubclassTypeArgumentMatches() {
-		this.contextRunner
-				.withUserConfiguration(ParameterizedWithCustomConfig.class, TypeArgumentsConditionWithValueConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("customExampleBean")));
-	}
-
-	@Test
-	void parameterizedContainerGenericWhenTypeArgumentNotMatches() {
-		this.contextRunner
-				.withUserConfiguration(GenericWithIntegerTypeArgumentsConfig.class,
-						TypeArgumentsConditionWithParameterizedContainerConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("genericIntegerExampleBean",
-						"parameterizedContainerGenericExampleBean")));
-	}
-
-	@Test
-	void parameterizedContainerGenericWhenTypeArgumentMatches() {
-		this.contextRunner
-				.withUserConfiguration(GenericWithStringTypeArgumentsConfig.class,
-						TypeArgumentsConditionWithParameterizedContainerConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("genericStringExampleBean")));
-	}
-
-	@Test
-	void parameterizedContainerGenericWhenSubclassTypeArgumentMatches() {
-		this.contextRunner
-				.withUserConfiguration(ParameterizedWithCustomConfig.class,
-						TypeArgumentsConditionWithParameterizedContainerConfig.class)
-				.run((context) -> assertThat(context).satisfies(exampleBeanRequirement("customExampleBean")));
-	}
-
 	private Consumer<ConfigurableApplicationContext> exampleBeanRequirement(String... names) {
 		return (context) -> {
 			String[] beans = context.getBeanNamesForType(ExampleBean.class);
@@ -418,8 +353,8 @@ class ConditionalOnMissingBeanTests {
 
 		@Bean
 		@ConditionalOnMissingBean(search = SearchStrategy.ANCESTORS)
-		ExampleBean<String> exampleBean2() {
-			return new ExampleBean<String>("test");
+		ExampleBean exampleBean2() {
+			return new ExampleBean("test");
 		}
 
 	}
@@ -451,7 +386,7 @@ class ConditionalOnMissingBeanTests {
 	static class FactoryBeanConfiguration {
 
 		@Bean
-		FactoryBean<ExampleBean<String>> exampleBeanFactoryBean() {
+		FactoryBean<ExampleBean> exampleBeanFactoryBean() {
 			return new ExampleFactoryBean("foo");
 		}
 
@@ -477,7 +412,7 @@ class ConditionalOnMissingBeanTests {
 	static class FactoryBeanWithBeanMethodArgumentsConfiguration {
 
 		@Bean
-		FactoryBean<ExampleBean<String>> exampleBeanFactoryBean(@Value("${theValue}") String value) {
+		FactoryBean<ExampleBean> exampleBeanFactoryBean(@Value("${theValue}") String value) {
 			return new ExampleFactoryBean(value);
 		}
 
@@ -568,8 +503,8 @@ class ConditionalOnMissingBeanTests {
 
 		@Bean
 		@ConditionalOnMissingBean(ExampleBean.class)
-		ExampleBean<String> createExampleBean() {
-			return new ExampleBean<>("direct");
+		ExampleBean createExampleBean() {
+			return new ExampleBean("direct");
 		}
 
 	}
@@ -579,8 +514,8 @@ class ConditionalOnMissingBeanTests {
 
 		@Bean
 		@ConditionalOnMissingBean(value = ExampleBean.class, ignored = CustomExampleBean.class)
-		ExampleBean<String> exampleBean() {
-			return new ExampleBean<>("test");
+		ExampleBean exampleBean() {
+			return new ExampleBean("test");
 		}
 
 	}
@@ -591,8 +526,8 @@ class ConditionalOnMissingBeanTests {
 		@Bean
 		@ConditionalOnMissingBean(value = ExampleBean.class,
 				ignoredType = "org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBeanTests$CustomExampleBean")
-		ExampleBean<String> exampleBean() {
-			return new ExampleBean<>("test");
+		ExampleBean exampleBean() {
+			return new ExampleBean("test");
 		}
 
 	}
@@ -666,8 +601,8 @@ class ConditionalOnMissingBeanTests {
 	static class ExampleBeanConfiguration {
 
 		@Bean
-		ExampleBean<String> exampleBean() {
-			return new ExampleBean<>("test");
+		ExampleBean exampleBean() {
+			return new ExampleBean("test");
 		}
 
 	}
@@ -677,21 +612,21 @@ class ConditionalOnMissingBeanTests {
 
 		@Bean
 		@ConditionalOnMissingBean
-		ExampleBean<String> exampleBean2() {
-			return new ExampleBean<>("test");
+		ExampleBean exampleBean2() {
+			return new ExampleBean("test");
 		}
 
 	}
 
-	static class ExampleFactoryBean implements FactoryBean<ExampleBean<String>> {
+	static class ExampleFactoryBean implements FactoryBean<ExampleBean> {
 
 		ExampleFactoryBean(String value) {
 			Assert.state(!value.contains("$"), "value should not contain '$'");
 		}
 
 		@Override
-		public ExampleBean<String> getObject() {
-			return new ExampleBean<>("fromFactory");
+		public ExampleBean getObject() {
+			return new ExampleBean("fromFactory");
 		}
 
 		@Override
@@ -713,8 +648,8 @@ class ConditionalOnMissingBeanTests {
 		}
 
 		@Override
-		public ExampleBean<String> getObject() {
-			return new ExampleBean<>("fromFactory");
+		public ExampleBean getObject() {
+			return new ExampleBean("fromFactory");
 		}
 
 		@Override
@@ -803,67 +738,23 @@ class ConditionalOnMissingBeanTests {
 
 	}
 
-	@Configuration(proxyBeanMethods = false)
-	static class GenericWithStringTypeArgumentsConfig {
-
-		@Bean
-		@ConditionalOnMissingBean
-		ExampleBean<String> genericStringExampleBean() {
-			return new ExampleBean<>("genericStringExampleBean");
-		}
-
-	}
-
-	@Configuration(proxyBeanMethods = false)
-	static class GenericWithIntegerTypeArgumentsConfig {
-
-		@Bean
-		@ConditionalOnMissingBean
-		ExampleBean<Integer> genericIntegerExampleBean() {
-			return new ExampleBean<>(1_000);
-		}
-
-	}
-
-	@Configuration(proxyBeanMethods = false)
-	static class TypeArgumentsConditionWithValueConfig {
-
-		@Bean
-		@ConditionalOnMissingBean(ExampleBean.class)
-		ExampleBean<String> genericStringWithValueExampleBean() {
-			return new ExampleBean<>("genericStringWithValueExampleBean");
-		}
-
-	}
-
-	@Configuration(proxyBeanMethods = false)
-	static class TypeArgumentsConditionWithParameterizedContainerConfig {
-
-		@Bean
-		@ConditionalOnMissingBean(parameterizedContainer = TestParameterizedContainer.class)
-		TestParameterizedContainer<ExampleBean<String>> parameterizedContainerGenericExampleBean() {
-			return new TestParameterizedContainer<>();
-		}
-
-	}
-
 	@TestAnnotation
-	static class ExampleBean<T> {
+	static class ExampleBean {
 
-		private final T value;
+		private String value;
 
-		ExampleBean(T value) {
+		ExampleBean(String value) {
 			this.value = value;
 		}
 
 		@Override
 		public String toString() {
-			return String.valueOf(this.value);
+			return this.value;
 		}
 
 	}
 
-	static class CustomExampleBean extends ExampleBean<String> {
+	static class CustomExampleBean extends ExampleBean {
 
 		CustomExampleBean() {
 			super("custom subclass");
@@ -871,7 +762,7 @@ class ConditionalOnMissingBeanTests {
 
 	}
 
-	static class OtherExampleBean extends ExampleBean<String> {
+	static class OtherExampleBean extends ExampleBean {
 
 		OtherExampleBean() {
 			super("other subclass");
